more wrap cases

+++ plugins/XPlayBuf/XPlayBuf.hpp
@@ -47,7 +47,7 @@ private:
 
     int32 m_guardFrame;
     uint32 m_numWriteChannels;
-    float m_fadeSamples;
+    int32 m_fadeSamples;
     float m_OneOverFadeSamples;
     float m_remainingFadeSamples;
     float m_argLoopStart;
